+++ config/index.js
@@ -11,9 +11,9 @@ module.exports = {
     assetsPublicPath: '/',
     proxyTable: {
       '/api': {
-          // target: 'http://192.168.1.124:8080',//辉哥后端接口地址  
+          target: 'http://192.168.1.124:8080',//辉哥后端接口地址  
           // target:'http://39.98.82.63:8080',
-          target: 'http://192.168.1.145:8080',//钟宇宏后端接口地址  
+          // target: 'http://192.168.1.145:8080',//钟宇宏后端接口地址  
           changeOrigin: true,//是否允许跨越
           pathRewrite: {
               '^/api': '/',//重写,
