+++ src/pages/examManage/examManage.vue
@@ -3,39 +3,21 @@
         <el-tabs type="border-card">
             <el-tab-pane>
                 <span slot="label"><i class="el-icon-date iconfont icon-kaochangguanli-"></i>考场管理</span>
-                <exam-roomList :SHOWTYPE="SHOW_TYPE" v-if="defaultValue === SHOW_TYPE.LIST"  @showDiffPage = 'showDiffPage'></exam-roomList>
-                <frozen-room  :SHOWTYPE="SHOW_TYPE" v-else-if="defaultValue === SHOW_TYPE.FROZEN"></frozen-room>
-                <add-room   :SHOWTYPE="SHOW_TYPE"  v-else="defaultValue === SHOW_TYPE.ADD"></add-room>
+                <examRoomListComponent></examRoomListComponent>
             </el-tab-pane> 
         </el-tabs>
     </div>
 </template>
 <script>
-import examRoomList from "./examRoomList";
-import frozenRoom from "./frozenRoom";
-import addRoom from "./addRoom";
-
+import examRoomListComponent from "./examRoomListComponent";
 export default {
   components:{
-    'exam-roomList':examRoomList,
-    'frozen-room':frozenRoom,
-    'add-room':addRoom
+      examRoomListComponent,
   },
   data(){
     return{
-      SHOW_TYPE: {
-        LIST: "list",
-        ADD: "add",
-        FROZEN: "frozen"
-      },
-      defaultValue: "list"      
     }
   },
-  methods:{
-    showDiffPage(val){
-      this.defaultValue = val;
-    }
-  }
 }
 </script>
 
@@ -51,7 +33,6 @@ export default {
       padding: 0;
     }
   }
-  
 }
 </style>
