+++ src/pages/standardManage/standardDetail/standardDetail.vue
@@ -7,85 +7,61 @@
             <div class="standardBody">
                 <div class="testLevel">
                     <span class="commontips">考试级别:</span>
-                    <el-select v-model="examLevel" placeholder="请选择" :disabled="disabled">
-                        <!-- <el-option
-                        v-for="item in options1"
-                        :key="item.value"
-                        :label="item.label"
-                        :value="item.value">
-                        </el-option> -->
-                    </el-select>
+                    <el-input v-model="examLevel"  disabled></el-input>
                 </div>
 
                 <div class="testLevel">
                     <span class="commontips">报考条件:</span>
-                    <el-select v-model="signRequirement" placeholder="请选择"  :disabled="disabled">
-                        <!-- <el-option
-                        v-for="item in options2"
-                        :key="item.value"
-                        :label="item.label"
-                        :value="item.value">
-                        </el-option> -->
+                    <el-select v-model="signRequirement" disabled>
+                      
                     </el-select>
                 </div>
 
-                
                 <div class="testLevel">
                     <span class="commontips">题库难度:</span>
-                    <el-select v-model="itemDifficulty" placeholder="请选择"  :disabled="disabled"></el-select>
-                    <!-- <el-input-number v-model="itemDifficulty" controls-position="right"  :disabled="disabled"></el-input-number> -->
+                    <div class="ajustBtn">
+                        <el-input v-model="examLevelName" disabled></el-input>
+                        <button class="addLevel" disabled>+</button>
+                        <button class="reduceLevel" disabled>—</button>
+                    </div>
                 </div>
 
                 <div class="knowledgeHierarchy">
                   <span>知识体系:</span> 
                   <p>
-                    <span v-for="(item,index) in knowledgeHierarchy" :key="index" >{{item}}道</span>
-                    <!-- <span>中盘2道</span>
-                    <span>布局2道</span>
-                    <span>官子2道</span>
-                    <span>死活2道</span>
-                    <span>对弈2道</span> -->
+                    <span v-for="(item,index) in knowledgeHierarchy" :key="index" >{{item.split('@')[0]}} {{item.split('@')[1]}}道</span>
                   </p>
                 </div>
 
                  <div class="testTime">
                   <span>考试时长:</span> 
-                  <p v-html='examLength'></p>
+                  <p>{{examLength}}分钟</p>
                 </div>
 
             </div>
 
                 <div class="examFee">
                     <span>考试服务费:（元）</span>
-                    <el-input v-model="examServiceFee" :disabled="disabled"></el-input>
+                    <el-input v-model.number="examServiceFee" disabled></el-input>
                 </div>
 
                 <div class="examFee">
                     <span>认证服务费:（元）</span>
-                    <el-input v-model="certificationServiceFee" :disabled="disabled"></el-input>
+                    <el-input v-model.number="certificationServiceFee" disabled></el-input>
                 </div>
 
-                <div class="examFee detailInfo">
-                    <span>详情介绍:</span>
+                <div class="examFee detailInfo"  v-if="false">
+                    <span v-if="false">详情介绍:</span>
                     <el-input
                       type="textarea"
+                      v-if="false"
                       :autosize="{ minRows: 4, maxRows: 4}"
                       maxlength="200"
+                      @input = "descInput"
                       v-model="details"
-                      :disabled="disabled"
                       >
                     </el-input>
-                    <p>已输入<span v-html="number"></span>个字,不得超过200字</p>
-                    <!-- <div style="text-align:center;width:100%;">
-                      <el-button type="primary" round 
-                        style="background:#1f91b5;
-                        width:145px;
-                        heiht:40px;
-                        font-size:16px;
-                        margin-top:38px;"
-                        @click="handleput"  
-                      >回到上一页</el-button>
-                    </div> -->
+                    <p v-if="false">已输入<span v-html="number"></span>个字,不得超过200字</p>
                 </div>
 
         </div>
@@ -105,57 +81,72 @@ export default {
   data() {
     return {
       titleTop: "考题标准详情",
+      dataType:1,
+      id:'',
       examLevel: null,
+      examLevelName:'',
       signRequirement: null,
-      itemDifficulty: null,
       examServiceFee: null,
       certificationServiceFee:null,
-      knowledgeHierarchy:null,
-      examLength:null,
-      disabled: true,
-      details: null,
-      number: 0
+      knowledgeHierarchy:[],
+      examLength:"0",
+      itemDifficulty:'',
+      details:null,
+      number:0,
+      examLevelList: [],
     };
   },
   methods: {
-    handleput() {}
+    descInput(){
+      details:null,
+      this.number = this.details.length
+    },
   },
-  created() {
-    let id = this.$route.query.id;
-    let params = new URLSearchParams();
-    params.append("id", id);
-    this.$http.get("/api/standard/standard_detail",{params}).then(res => {
-        console.log(res.data.data);
-        let rst = res.data.data;
-        //处理数据
-        this.examLevel = rst.examLevel +'级'
-        this.signRequirement = rst.signRequirement+'级'
-        this.itemDifficulty = rst.itemDifficulty+'级'
-        this.certificationServiceFee = rst.certificationServiceFee
-        this.examServiceFee = rst.examServiceFee
-        this.examLength = rst.examLength
-        this.details = rst.details
-        this.knowledgeHierarchy = rst.knowledgeHierarchy.split(',').splice(0,5)
-        this.number = rst.details.length
-        console.log(this.knowledgeHierarchy)
-        })
+  created(){
+      //获取考试级别
+      let params1 ={
+        // "levelId":1
+      }
+      this.$http.get('/api/standard/standard_level_list',{params1}).then( res =>{
+        if(res.data.code === 0){
+          console.log(res.data.data.levelList)
+          this.examLevelList = res.data.data.levelList;
+        }
+      }).then( ()=>{
+          let id = this.$route.query.id;
+          this.id = id;
+          let url = '/api/standard/standard_detail?id='+id
+          this.$http.get(url).then(res => {
+            console.log(res.data.data,'8888888');
+            let rst = res.data.data;
+            //处理数据
+            this.examLevel = this.examLevelList.filter( item => item.id == rst.examLevel)[0].levelName;
+            this.signRequirement = this.examLevelList.filter( item => item.id == rst.signRequirement)[0].levelName;
+            this.itemDifficulty = this.examLevelList.filter( item => item.id == rst.itemDifficulty)[0].levelName;
+            this.certificationServiceFee = rst.certificationServiceFee
+            this.examServiceFee = rst.examServiceFee
+            this.examLength = rst.examLength
+            this.knowledgeHierarchy = rst.knowledgeHierarchy.split(',')
+            this.itemDifficulty = rst.itemDifficulty;
+            this.examLevelName = this.examLevelList.filter( item => item.id == rst.itemDifficulty)[0].levelName;
+            // this.details = rst.details
+            // this.number = rst.details.length    
+          })
+      })
   }
-  
 };
 </script>
-
-
 <style rel='stylesheet/scss' lang="scss">
-.el-tabs--border-card > .el-tabs__content {
+.el-tabs--border-card>.el-tabs__content{
   padding: 0;
 }
-#standardTemplateBox {
+#standardTemplateBox{
   width: 100%;
   background: #f0f0f0;
 }
-.standardTemplate {
-  width: 1042px;
-  height: auto;
+.standardTemplate{
+    width: 1042px;
+    height: auto;
 }
 .addStandard {
   width: 952px;
@@ -186,14 +177,57 @@ export default {
         padding-bottom: 18px;
         padding-left: 7px;
       }
+      &>.ajustBtn{
+          width: 160px;
+          height: 40px;
+          position: relative;
+          & /deep/ .el-input__inner{
+            width: 134px;
+            border-radius: 4px 0 0 4px;
+          }
+          &>.addLevel{
+            width: 26px;
+            height: 20px;
+            border: none;
+            outline: none;
+            position: absolute;
+            right:0px;
+            top: 0px;
+            background: #ffffff;
+            border: 1px solid #DCDFE6;
+            border-radius:0 4px 0  0;
+            text-align: center;
+            font-size: 16px;
+            border-left:none;
+            line-height: 18px;
+          } 
+          &>.reduceLevel{
+            width: 26px;
+            height: 20px;
+            border: none;
+            outline: none;
+            position: absolute;
+            right: 0px;
+            top: 20px;
+            border-radius:0 0 4px  0;
+            text-align: center;
+            font-size: 12px;
+            background: #ffffff;
+            border: 1px solid #DCDFE6;
+            border-top:none;
+            border-left:none;
+            line-height: 15px;
+          }
+      }
+     
     }
-
+   
     .knowledgeHierarchy {
       width: 100%;
       height: 121px;
       float: left;
       margin-bottom: 36px;
-      & > span {
+      &>span{
         width: 164px;
         height: 17px;
         display: block;
@@ -205,22 +239,24 @@ export default {
         padding-left: 7px;
       }
       & > p {
+        width: calc( 100% - 7px);
         padding-left: 7px;
         height: 70px;
         & > span {
           display: block;
           float: left;
-          width: 112px;
+          width: 140px;
           font-size: 16px;
           height: 70px;
           line-height: 70px;
+          margin-right: 10px;
         }
       }
     }
     .testTime {
       float: left;
       height: 53px;
-      & > span {
+      &>span{
         width: 164px;
         height: 17px;
         display: block;
@@ -233,7 +269,7 @@ export default {
       & > p {
         padding-left: 7px;
         width: 112px;
-
+        
         font-size: 16px;
       }
     }
@@ -252,16 +288,16 @@ export default {
     }
   }
 
-  .detailInfo {
+  .detailInfo{
     width: 100%;
     height: 125px;
     margin-bottom: 34px;
-    & > p {
+    &>p{
       float: right;
       height: 42px;
       line-height: 42px;
       font-size: 14px;
-      color: #1f91b5;
+      color:#1f91b5;
     }
   }
 
@@ -273,6 +309,8 @@ export default {
     color: #000;
   }
 }
+
+
 </style>
