+++ src/router/indexo.js
@@ -21,17 +21,7 @@ export default new Router({
           component: () => import('@/pages/examManage/examManage'),
         },
         {
-          path: '/examManage/addRoom',
-          name: 'addRoom',
-          component: () => import('@/pages/examManage/addRoom'),
-        },
-        {
-          path: '/examManage/frozenRoom',
-          name: 'frozenRoom',
-          component: () => import('@/pages/examManage/frozenRoom'),
-        },
-        {
-          path: '/examManage/examRoomDetail',
+          path: '/examRoomDetail',
           name: 'examRoomDetail',
           component: () => import('@/pages/examManage/examRoomDetail'),
         },
@@ -82,26 +72,28 @@ export default new Router({
           path: '/planManage', //考试计划
           name: 'planManage',
           component: () => import('@/pages/planManage/planManage'),
-        },
-        {
-          path: '/planManage/addPlan', //考试计划 
-          name: 'addPlan',
-          component: () => import('@/pages/planManage/addPlan')
-        },
-        {
-          path: '/planManage/editPlan', //考试计划 
-          name: 'editPlan',
-          component: () => import('@/pages/planManage/editPlan')
+          children:[
+            {
+              path: '/plansList', //考试计划 
+              name: 'plansList',
+              component: () => import('@/pages/planManage/plansList')
+            },
+            {
+              path: '/addPlan', //考试计划 
+              name: 'addPlan',
+              component: () => import('@/pages/planManage/addPlan')
+            },
+            {
+              path: '/editPlan', //考试计划 
+              name: 'editPlan',
+              component: () => import('@/pages/planManage/editPlan')
+            }
+          ]
         },
         {
           path: '/orderManage', //订单管理
           name: 'orderManage',
           component: () => import('@/pages/orderManage/orderManage'),
-        },
-        {
-          path: '/orderManage/orderDetail', //订单管理
-          name: 'orderDetail',
-          component: () => import('@/pages/orderManage/orderDetail'),
         },
         {
           path: '/examStatus',
